+++ examples/runner-default/main.tf
@@ -9,7 +9,7 @@ data "aws_security_group" "default" {
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "2.70"
+  version = "3.19.0"
 
   name = "vpc-${var.environment}"
   cidr = "10.0.0.0/16"
@@ -20,7 +20,24 @@ module "vpc" {
 
   enable_nat_gateway = true
   single_nat_gateway = true
-  enable_s3_endpoint = true
+
+  tags = {
+    Environment = var.environment
+  }
+}
+
+module "vpc_endpoints" {
+  source  = "terraform-aws-modules/vpc/aws//modules/vpc-endpoints"
+  version = "3.19.0"
+
+  vpc_id = module.vpc.vpc_id
+
+  endpoints = {
+    s3 = {
+      service = "s3"
+      tags    = { Name = "s3-vpc-endpoint" }
+    }
+  }
 
   tags = {
     Environment = var.environment
@@ -125,6 +142,13 @@ module "runner" {
   #   command    = ["--registry-mirror", "https://mirror.gcr.io"]
   #   entrypoint = ["dockerd-entrypoint.sh"]
   # }]
+
+
+  # Example how to configure runners, to utilize EC2 user-data feature
+  # example template, creates (configurable) swap file for the runner
+  # runners_userdata = templatefile("${path.module}/../../templates/swap.tpl", {
+  #   swap_size = "512"
+  # })
 }
 
 resource "null_resource" "cancel_spot_requests" {